Enhancements to Diagnostics
Add E2E guide to run locally (#337)

* Add E2E guide to run locally

* Add export command depending on OS

* Update CONTRIBUTING.md

Co-authored-by: Shivank Dubey <110578314+dubeyShivank@users.noreply.github.com>

---------

Co-authored-by: Shivank Dubey <110578314+dubeyShivank@users.noreply.github.com>

Updated links and descriptions in the README (#338)

* Added links to npm and release shields

Pointed shields/badges to the npm page and releases pages instead of loading the shield itself as an image.

* Update table of contents

The link to the React-Native section was broken.
Removed links to overall page and table of contents header.

* added multiview demo link

The link to the multiview demo was missing and the introduction to this section needed an update.

* updated basic usage

The old url redirected, but updated to point directly to the streaming dashboard. Also made the callouts for setting the token and stream details before the code sample instead of after it.

* updated documentation links

Added links to the documentation overview, getting started, and the SDK reference under a Documentation header.

---------

Co-authored-by: Shivank Dubey

Typescript definitions for diagnose (#345)

* Add typescript definition for diagnose

* Add in changeset

* Remove changeset

* Add type definition for PeerConnectionStats

* Add remaining type definitions

Fix types & add outbound stats to CMCD-like structure

Add changeset